Добавляет новые компоненты
