+++ project/src/components/app/app.tsx
@@ -1,11 +1,20 @@
-import MainScreen from '../../screens/main-screen/main-screen';
+import Property from '../../screens/property/property';
 
-type Cards = {
-  cards: number[]
-};
+// type Cards = {
+//   cards: number[]
+// };
 
-function App({cards} : Cards): JSX.Element {
-  return <MainScreen cards={cards}/>;
+function App(): JSX.Element {
+  return <Property/>;
 }
+// import MainScreen from '../../screens/main-screen/main-screen';
+
+// type Cards = {
+//   cards: number[]
+// };
+
+// function App({cards} : Cards): JSX.Element {
+//   return <MainScreen cards={cards}/>;
+// }
 
 export default App;
