+++ project/src/components/favorite-locations-list/favorite-locations-list.tsx
@@ -1,0 +1,12 @@
+import FavoriteLocationItems from '../favorite-location-items/favorite-location-items';
+
+function FavoriteLocationsList(): JSX.Element {
+  return (
+    <ul className="favorites__list">
+      <FavoriteLocationItems/>
+      <FavoriteLocationItems/>
+    </ul>
+  );
+}
+
+export default FavoriteLocationsList;
