+++ project/src/components/near-places-list/near-places-list.tsx
@@ -1,0 +1,13 @@
+import NearPlaceCard from '../near-place-card/near-place-card';
+
+function NearPlacesList(): JSX.Element {
+  return (
+    <div className="near-places__list places__list">
+      <NearPlaceCard/>
+      <NearPlaceCard/>
+      <NearPlaceCard/>
+    </div>
+  );
+}
+
+export default NearPlacesList;
