+++ project/src/components/property-features-list/property-features-list.tsx
@@ -1,0 +1,17 @@
+function PropertyFeaturesList(): JSX.Element {
+  return (
+    <ul className="property__features">
+      <li className="property__feature property__feature--entire">
+        Apartment
+      </li>
+      <li className="property__feature property__feature--bedrooms">
+        3 Bedrooms
+      </li>
+      <li className="property__feature property__feature--adults">
+        Max 4 adults
+      </li>
+    </ul>
+  );
+}
+
+export default PropertyFeaturesList;
