+++ project/src/components/property-inside-list/property-inside-list.tsx
@@ -1,0 +1,38 @@
+function PropertyInsideList(): JSX.Element {
+  return (
+    <ul className="property__inside-list">
+      <li className="property__inside-item">
+        Wi-Fi
+      </li>
+      <li className="property__inside-item">
+        Washing machine
+      </li>
+      <li className="property__inside-item">
+        Towels
+      </li>
+      <li className="property__inside-item">
+        Heating
+      </li>
+      <li className="property__inside-item">
+        Coffee machine
+      </li>
+      <li className="property__inside-item">
+        Baby seat
+      </li>
+      <li className="property__inside-item">
+        Kitchen
+      </li>
+      <li className="property__inside-item">
+        Dishwasher
+      </li>
+      <li className="property__inside-item">
+        Cabel TV
+      </li>
+      <li className="property__inside-item">
+        Fridge
+      </li>
+    </ul>
+  );
+}
+
+export default PropertyInsideList;
