+++ project/src/index.tsx
@@ -2,11 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app/app';
 
-const CARDS_AMOUNT = 5;
-const cards = new Array(CARDS_AMOUNT).fill('').map((element, index) => element = index);
+// const CARDS_AMOUNT = 5;
+// const cards = new Array(CARDS_AMOUNT).fill('').map((element, index) => element = index);
 
 ReactDOM.render(
   <React.StrictMode>
-    <App cards={cards}/>
+    {/* <App cards={cards}/> */}
+    <App/>
   </React.StrictMode>,
   document.getElementById('root'));
